requirements.txt
bs4
python-telegram-bot
psutil
pysocks
emoji